+++ components/admin/pages/form/PagesForm.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import cx from 'classnames';
 
 // Services
 import { updatePage, createPage, fetchPage } from 'services/pages';
@@ -88,7 +89,7 @@ class PagesForm extends React.Component {
                 this.setState({ submitting: false });
                 toastr.error(`There was an error updating the page: ${id}" - "${form.title}`, err);
               });
-          // Create page
+            // Create page
           } else {
             createPage(form, authorization)
               .then((data) => {
@@ -144,27 +145,35 @@ class PagesForm extends React.Component {
   }
 
   render() {
+    const { form, loading, id, step, stepLength, submitting } = this.state;
     return (
-      <form className="c-form" onSubmit={this.onSubmit} noValidate>
-        <Spinner isLoading={this.state.loading} className="-light" />
+      <form
+        className={cx({
+          'c-form': true,
+          '-disabled': form.env && !process.env.NEXT_PUBLIC_ENVS_EDIT.includes(form.env),
+        })}
+        onSubmit={this.onSubmit}
+        noValidate
+      >
+        <Spinner isLoading={loading} className="-light" />
 
-        {(this.state.step === 1 && !this.state.loading)
+        {(step === 1 && !loading)
           && (
-          <Step1
-            onChange={(value) => this.onChange(value)}
-            form={this.state.form}
-            id={this.state.id}
-          />
+            <Step1
+              onChange={(value) => this.onChange(value)}
+              form={form}
+              id={id}
+            />
           )}
 
-        {!this.state.loading
+        {!loading
           && (
-          <Navigation
-            step={this.state.step}
-            stepLength={this.state.stepLength}
-            submitting={this.state.submitting}
-            onStepChange={this.onStepChange}
-          />
+            <Navigation
+              step={step}
+              stepLength={stepLength}
+              submitting={submitting}
+              onStepChange={this.onStepChange}
+            />
           )}
       </form>
     );