+++ components/admin/pages/table/PagesTable.js
@@ -98,6 +98,9 @@ class PagesTable extends PureComponent {
     const { filteredPages } = this.props;
     const { pagination } = this.state;
 
+    const pagesWithDisabled = filteredPages.map((p) => (
+      { ...p, disabled: !process.env.NEXT_PUBLIC_ENVS_EDIT.includes(p.env) }));
+
     return (
       <div className="c-pages-table">
         <Spinner className="-light" isLoading={this.props.loading} />
@@ -143,7 +146,7 @@ class PagesTable extends PureComponent {
               value: 1,
             }}
             filters={false}
-            data={filteredPages}
+            data={pagesWithDisabled}
             manualPagination
             onChangePage={this.onChangePage}
             onRowDelete={() => this.props.getPages()}