+++ components/datasets/form/steps/component.js
@@ -1,6 +1,7 @@
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
 import compact from 'lodash/compact';
+import cx from 'classnames';
 
 // Constants
 import {
@@ -33,11 +34,13 @@ class Step1 extends PureComponent {
     onChange: PropTypes.func.isRequired,
     sortedLayers: PropTypes.array,
     user: PropTypes.object.isRequired,
+    disabled: PropTypes.bool,
   };
 
   static defaultProps = {
     dataDataset: null,
     sortedLayers: [],
+    disabled: false,
   };
 
   state = {
@@ -157,7 +160,7 @@ class Step1 extends PureComponent {
 
   render() {
     const {
-      user, columns, loadingColumns, basic, sortedLayers, authorization,
+      user, columns, loadingColumns, basic, sortedLayers, authorization, disabled, onChange,
     } = this.props;
     const { dataset, subscribableSelected } = this.state;
     const { provider, columnFields, application } = this.state.form;
@@ -180,29 +183,24 @@ class Step1 extends PureComponent {
     const columnFieldsOptions = (columnFields || []).map((f) => ({ label: f, value: f }));
 
     return (
-      <div>
+      <div className={cx({ '-disabled': disabled })}>
         <fieldset className="c-field-container">
           {user.role === 'ADMIN' && !basic && (
             <Field
               ref={(c) => {
                 if (c) FORM_ELEMENTS.elements.env = c;
               }}
-              hint={
-                'Choose "preproduction" to see this dataset it only as admin, "production" option will show it in public site.'
-              }
               className="-fluid"
               options={[
-                { label: 'Pre-production', value: 'preproduction' },
+                { label: 'Staging', value: 'staging' },
+                { label: 'Preproduction', value: 'preproduction' },
                 { label: 'Production', value: 'production' },
               ]}
-              onChange={(value) => this.props.onChange({ env: value })}
+              onChange={(value) => onChange({ env: value })}
               properties={{
                 name: 'env',
                 label: 'Environment',
-                placeholder: 'Type the columns...',
-                noResultsText: 'Please, type the name of the columns and press enter',
-                promptTextCreator: (label) => `The name of the column is "${label}"`,
-                default: 'preproduction',
+                default: process.env.NEXT_PUBLIC_API_ENV,
                 value: this.props.form.env,
               }}
             >