+++ layout/layout/ocean-watch/storytelling/story-step/component.jsx
@@ -26,16 +26,10 @@ import {
   getRWAdapter,
 } from 'utils/widget-editor';
 
-import {
-  isStagingAPI,
-} from 'utils/api';
-
 // styles
 import './styles.scss';
 
 const WidgetShareModal = dynamic(() => import('../../../../../components/widgets/share-modal'), { ssr: false });
-
-const isStaging = isStagingAPI();
 
 function renderWidget({
   id: widgetId,
@@ -55,6 +49,7 @@ function renderWidget({
           style={{
             height: 450,
             borderRadius: 4,
+            color: '#393f44',
           }}
         />
       )}
@@ -163,7 +158,7 @@ export default function StoryStep({
   );
 
   const widgetParams = useMemo(() => ({
-    geostore_env: isStaging ? 'geostore_staging' : 'geostore_prod',
+    geostore_env: 'geostore_prod',
     ...geostore && { geostore_id: geostore },
   }), [geostore]);